// ui/src/views/dashboard/components/PanelEditor/types.ts
export interface PanelEditorTab {
    id: string;
    text: string;
    active: boolean;
    icon: string;
  }
  
  export enum PanelEditorTabId {
    Query = 'query',
    Transform = 'transform',
    Visualize = 'visualize',
    Alert = 'alert',
  }
  
  export enum DisplayMode {
    Fill = 0,
    Fit = 1,
    Exact = 2,
  }
  
  export const displayModes = [
    { value: DisplayMode.Fill, label: 'Fill', description: 'Use all available space' },
    { value: DisplayMode.Fit, label: 'Fit', description: 'Fit in the space keeping ratio' },
    { value: DisplayMode.Exact, label: 'Exact', description: 'Same size as the dashboard' },
  ];
  

export interface PanelEditorUIState {
    /* Visualization options pane visibility */
    isPanelOptionsVisible: boolean;
    /* Pixels or percentage */
    rightPaneSize: number | string;
    /* Pixels or percentage */
    topPaneSize: number | string;
    /* Visualization size mode */
    mode: DisplayMode;
  }